be/requirements.txt
django >= 5.1.1
djangorestframework
django-cors-headers
django-filter
psycopg2-binary
pillow
django-extensions >= 3.2.3